fixes of several errors and warnings indicated by LGTM

+++ src/data_io/MNISTImageImporter.hpp
@@ -79,12 +79,12 @@ private:
 	/*!
 	 * Width of MNIST image.
 	 */
-	int image_width;
+	size_t image_width;
 
 	/*!
 	 * Height of MNIST image.
 	 */
-	int image_height;
+	size_t image_height;
 
 	/*!
 	 * Property: directory/Name of file containing images (binary datafile).
